charmversion str addition to archive
