+++ charmdir.go
@@ -7,6 +7,7 @@ import (
 	"archive/zip"
 	"errors"
 	"fmt"
+	"github.com/juju/utils"
 	"io"
 	"os"
 	"path/filepath"
@@ -172,10 +173,15 @@ func resolveSymlinkedRoot(rootPath string) (string, error) {
 // ArchiveTo creates a charm file from the charm expanded in dir.
 // By convention a charm archive should have a ".charm" suffix.
 func (dir *CharmDir) ArchiveTo(w io.Writer) error {
-	return writeArchive(w, dir.Path, dir.revision, dir.Meta().Hooks())
+	versionString, err := MaybeCreateGenerateString(dir.Path)
+	if err != nil {
+		logger.Debugf("version file creation/overwrite failed : %v", err)
+		versionString = ""
+	}
+	return writeArchive(w, dir.Path, dir.revision, versionString, dir.Meta().Hooks())
 }
 
-func writeArchive(w io.Writer, path string, revision int, hooks map[string]bool) error {
+func writeArchive(w io.Writer, path string, revision int, versionString string, hooks map[string]bool) error {
 	zipw := zip.NewWriter(w)
 	defer zipw.Close()
 
@@ -188,6 +194,9 @@ func writeArchive(w io.Writer, path string, revision int, hooks map[string]bool)
 	zp := zipPacker{zipw, rootPath, hooks}
 	if revision != -1 {
 		zp.AddRevision(revision)
+	}
+	if versionString != "" {
+		zp.AddVersion(versionString)
 	}
 	return filepath.Walk(rootPath, zp.WalkFunc())
 }
@@ -210,6 +219,16 @@ func (zp *zipPacker) AddRevision(revision int) error {
 	w, err := zp.CreateHeader(h)
 	if err == nil {
 		_, err = w.Write([]byte(strconv.Itoa(revision)))
+	}
+	return err
+}
+
+func (zp *zipPacker) AddVersion(version string) error {
+	h := &zip.FileHeader{Name: "version"}
+	h.SetMode(syscall.S_IFREG | 0644)
+	w, err := zp.CreateHeader(h)
+	if err == nil {
+		_, err = w.Write([]byte(version))
 	}
 	return err
 }
@@ -315,4 +334,23 @@ func checkFileType(path string, mode os.FileMode) error {
 		e = "file is a device: %q"
 	}
 	return fmt.Errorf(e, path)
+}
+
+// MaybeCreateVersionFile creates/overwrite charm version file.
+func MaybeCreateGenerateString(path string) (string, error) {
+	var cmdArgs []string
+	// Verify that it is revision control directory.
+	if _, err := os.Stat(filepath.Join(path, ".hg")); err == nil {
+		cmdArgs = []string{"hg", "id", "-n"}
+	} else if _, err = os.Stat(filepath.Join(path, ".bzr")); err == nil {
+		cmdArgs = []string{"bzr", "version-info"}
+	} else if _, err = os.Stat(filepath.Join(path, ".git")); err == nil {
+		cmdArgs = []string{"git", "describe", "--dirty"}
+	} else {
+		logger.Debugf("charm is not in revision control directory")
+		// Return empty string.
+		return "", nil
+	}
+
+	return utils.RunCommand(cmdArgs[0], cmdArgs[1:]...)
 }
